Set idle timer disabled when waiting for a live game to start
Before this change the app would let the device go to sleep while waiting for a game. Then you unlock the screen, the app reconnects and your game request is gone and you have to request a new one. And the idle timer was only being disabled when you open the live game. Now the idle timer is disabled whenever you have a live game or an outbound live challenge.